backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "slowapi>=0.1.9",
    "supabase>=2.27.1",
    "uvicorn>=0.40.0",
]
